Add mobile menu and fix some mobile stylings
